size

+++ 0x00-hello_world/6-size.c
@@ -15,10 +15,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of char: %ld byte(s)\n", (unsigned long)sizeof(a));
-	printf("Size of an int: %ld byte(s)\n", (unsigned long)sizeof(b));
-	printf("Size of a long int: %ld byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of long long int: %ld byte(s)\n", (unsigned long)sizeof(d));
-	printf("Size of a float: %ld byte(s)\n", (unsigned long)sizeof(e));
+	printf("Size of char: %ld byte(s)\n", sizeof (a));
+	printf("Size of an int: %ld byte(s)\n", sizeof(b));
+	printf("Size of a long int: %ld byte(s)\n", sizeof(c));
+	printf("Size of long long int: %ld byte(s)\n", sizeof(d));
+	printf("Size of a float: %ld byte(s)\n", sizeof(e));
 	return (0);
 }
